requirements.txt

---

### 📦 `requirements.txt`

```txt
python-telegram-bot==20.8
python-dotenv==1.0.1
